ai-agents-service/requirements-production.txt
# Production Dependencies for PM Simulator v2.0
# Install with: pip install -r requirements-production.txt

# Core Flask
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0

# Async Task Queue
celery[redis]>=5.3.0
redis>=4.6.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# HTTP Client for LLM APIs
aiohttp>=3.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Authentication
PyJWT>=2.8.0

# Rate Limiting
Flask-Limiter>=3.5.0

# Validation & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Monitoring
prometheus-flask-exporter>=0.22.0
prometheus-client>=0.18.0
# sentry-sdk[flask]>=1.28.0

# Production Server
gunicorn>=21.2.0

# Development (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.4.0
